(REF) SnapshotUpgradeTest - Extract CivixSnapshotUpgradeTestTrait and RunMethodsTrait
