--- src/CRM/CivixBundle/RunsIfException.php
@@ -1,6 +1,0 @@
-<?php
-
-namespace CRM\CivixBundle;
-
-class RunsIfException extends \RuntimeException {
-}
