+++ src/CRM/CivixBundle/SkippedMethodException.php
@@ -1,0 +1,6 @@
+<?php
+
+namespace CRM\CivixBundle;
+
+class SkippedMethodException extends \RuntimeException {
+}
